[TXT bug] Refactor to a shared module
