+++ app/models/g_suite.rb
@@ -1,4 +1,6 @@
 class GSuite < ApplicationRecord
+	include Shared::Domain
+
   has_one :application, class_name: 'GSuiteApplication', required: true
   has_many :accounts, class_name: 'GSuiteAccount'
   belongs_to :event
@@ -26,10 +28,5 @@ class GSuite < ApplicationRecord
 
   def set_application
     self.application = GSuiteApplication.find_by(domain: self.domain)
-  end
-
-  def domain_without_protocol
-    bad = ['http', ':', '/'].any? { |s| domain.to_s.include? s }
-    errors.add(:domain, 'shouldn’t include http(s):// or ending /') if bad
   end
 end
