+++ app/models/g_suite_application.rb
@@ -1,4 +1,5 @@
 class GSuiteApplication < ApplicationRecord
+	include Shared::Domain
   include Rejectable
 
   paginates_per 50
@@ -35,22 +36,5 @@ class GSuiteApplication < ApplicationRecord
 
   def under_review?
     rejected_at.nil? && canceled_at.nil? && accepted_at.nil?
-  end
-
-  private
-
-  def domain_without_protocol
-    bad = ['http', ':', '/'].any? { |s| domain.to_s.include? s }
-    errors.add(:domain, 'shouldn’t include http(s):// or ending /') if bad
-  end
-
-  def domain_not_email
-    errors.add(:domain, 'shouldn’t be an email address') if domain.to_s.include? '@'
-  end
-
-  def domain_is_lowercase
-    return if domain.to_s.downcase == domain
-
-    errors.add(:domain, 'must be all lowercase')
   end
 end
